git-svn-id: svn://192.168.2.130:8807/sd/java/financialSystem@60830 ced6f03c-c1bb-49ab-ae0c-4da96410de7b

+++ src/main/java/cn/financial/service/impl/UserServiceImpl.java
@@ -1,12 +1,12 @@
 package cn.financial.service.impl;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
@@ -18,6 +18,7 @@ import cn.financial.dao.UserRoleDAO;
 import cn.financial.model.User;
 import cn.financial.model.UserRole;
 import cn.financial.service.UserService;
+import cn.financial.util.UuidUtil;
 import cn.financial.util.shiro.PasswordHelper;
 
 
@@ -36,14 +37,20 @@ public class UserServiceImpl implements  UserService{
      * 查询全部/多条件查询用户列表
      */
     @Override
-    public java.util.List<User> listUser(Map<Object, Object> map) {
+    public List<User> listUser(Map<Object, Object> map) {
+        if(map.get("pageSize") == null || map.get("pageSize").equals("")){
+            map.put("pageSize",10);
+        }
+        if(map.get("page") == null || map.get("page").equals("")){
+            map.put("start", 0);
+        }
         return userDao.listUser(map);
     }
     /**
      * 查询全部/多条件查询用户列表  总条数
      */
     @Override
-    public java.util.List<User> listUserCount(Map<Object, Object> map) {
+    public List<User> listUserCount(Map<Object, Object> map) {
         return userDao.listUserCount(map);
     }
     /**
@@ -51,6 +58,12 @@ public class UserServiceImpl implements  UserService{
      */
     @Override
     public Integer countUserName(String name, String jobNumber) {
+        if(name == null || name.equals("")){
+            return -1;
+        }
+        if(jobNumber == null || jobNumber.equals("")){
+            return -2;
+        }
         return userDao.countUserName(name, jobNumber);
     }
     /**
@@ -58,6 +71,9 @@ public class UserServiceImpl implements  UserService{
      */
     @Override
     public User getUserById(String userId){
+        if(userId == null || userId.equals("")){
+          return null;  
+        }
         return userDao.getUserById(userId);
     }
     /**
@@ -72,18 +88,29 @@ public class UserServiceImpl implements  UserService{
      */
     @Override
     public Integer insertUser(User user) {
-        //加密密码
-        passwordHelper.encryptPassword(user);
-        return userDao.insertUser(user);
+        if(user.getRealName() == null || user.getRealName().equals("")){
+            return -1;
+        }else{
+          //加密密码
+            passwordHelper.encryptPassword(user);
+            return userDao.insertUser(user); 
+        }
     }
     /**
      * 修改用户
      */
     @Override
     public Integer updateUser(User user) {
-        //加密密码
-        if(user.getPwd()!=null && !user.getPwd().isEmpty()){
-            passwordHelper.encryptPassword(user);
+        if(user.getId() == null || user.getId().equals("")){//用户id
+            return -1;
+        }
+        if(user.getPwd()!=null && !user.getPwd().equals("")){
+            if(!user.getPwd().matches(User.REGEX)){//密码规则校验
+                return -2;
+            } 
+            user.setSalt(UuidUtil.getUUID());
+            user.setExpreTime(UuidUtil.expreTime());//添加密码到期时间
+            passwordHelper.encryptPassword(user);//加密密码
         }
         return userDao.updateUser(user);
     }
@@ -92,6 +119,9 @@ public class UserServiceImpl implements  UserService{
      */
     @Override
     public Integer deleteUser(String userId){
+        if(userId == null || userId.equals("")){
+            return -1;
+        }
         return userDao.deleteUser(userId);
     }
     //角色
@@ -134,7 +164,7 @@ public class UserServiceImpl implements  UserService{
         if(json.containsKey("children") && CollectionUtils.isNotEmpty(json.getJSONArray("children"))){
             for (Object item : json.getJSONArray("children")) {
                 JSONObject  itemChildren=(JSONObject)JSONObject.toJSON(item);
-                if(itemChildren.containsKey("orgType") && org.apache.commons.lang.StringUtils.isNotEmpty(itemChildren.getString("orgType"))){
+                if(itemChildren.containsKey("orgType") && StringUtils.isNotEmpty(itemChildren.getString("orgType"))){
                     permissions.add(itemChildren.getString("orgType"));
                     System.out.println("权限-----------"+itemChildren.getString("orgType"));
                 }
